+++ examples/resources/artifactory_package_cleanup_policy/import.sh
@@ -1,0 +1,1 @@
+terraform import artifactory_package_cleanup_policy.my-cleanup-policy my-policy