+++ pkg/artifactory/resource/configuration/resource_artifactory_repository_layout.go
@@ -20,7 +20,9 @@ import (
 )
 
 func NewRepositoryLayoutResource() resource.Resource {
-	return &RepositoryLayoutResource{}
+	return &RepositoryLayoutResource{
+		TypeName: "artifactory_repository_layout",
+	}
 }
 
 type RepositoryLayoutResource struct {
@@ -55,8 +57,7 @@ type RepositoryLayoutsAPIModel struct {
 }
 
 func (r *RepositoryLayoutResource) Metadata(ctx context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
-	resp.TypeName = req.ProviderTypeName + "_repository_layout"
-	r.TypeName = resp.TypeName
+	resp.TypeName = r.TypeName
 }
 
 func (r *RepositoryLayoutResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {